+++ modules/error.py
@@ -12,27 +12,47 @@ Exit code contexts:
     51-60: ZIP creation-specific
     61-70: Exclusion-specific
     71-80: Inclusion-specific
+    81-90: pack.mcmeta-specific
 """
 
 # Revision of this module:
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 
 import logging as log
+import traceback
 from modules.stats import codes
-from modules.arguments import get_args
+from modules.arguments import args
+
+def handle_log(header, err_type, messages):
+    # Add stacktrace if debug flag is set
+    if args.debug:
+        for line in traceback.format_exc().splitlines():
+            log.error(line)
+
+        log.error(f"{header} ({err_type}): {messages[0]}")
+        # Each extra message gets its own line
+        for message in messages[1:]:
+            log.error(message)
+
+        log.error("") # Separator
+    else:
+        log.error(f"Error ({err_type}): {messages[0]}")
+        for message in messages[1:]:
+            log.error(message)
 
 def record_warn(warn_type, message):
     codes['warning'][warn_type] += 1
-    log.warning(f"Warning ({warn_type}): {message}")
+    handle_log("Warning", warn_type, messages)
 
-def record_err(code, err_type, message):
+def record_err(code, err_type, *messages):
     codes['error'][err_type] += 1
-    log.error(f"Error ({err_type}): {message}")
+    handle_log("Error", err_type, messages)
     # Check exit_error flag
-    if get_args().exit_error:
+    if args.exit_error:
+        log.trace("exit_error was set, triggering exit")
         exit(code)
 
-def record_crit(code, crit_type, message):
+def record_crit(code, crit_type, *messages):
     codes['critical'][crit_type] += 1
-    log.critical(f"Fatal ({crit_type}): {message}")
+    handle_log("Fatal", crit_type, messages)
     exit(code)